# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ThreatLens AI API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get Me",
        "operationId": "get_me_api_v1_users_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "Update Me",
        "operationId": "update_me_api_v1_users_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/v1/threats/report": {
      "get": {
        "tags": [
          "threats"
        ],
        "summary": "Download Monitoring Report",
        "description": "Full monitoring snapshot, plus detections matching the Threat Monitor filters.",
        "operationId": "download_monitoring_report_api_v1_threats_report_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "window",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(24h|7d|30d)$",
              "default": "24h",
              "title": "Window"
            }
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(low|medium|high)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Level"
            }
          },
          {
            "name": "verdict",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(malicious|suspicious|benign)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Verdict"
            }
          },
          {
            "name": "family",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 120
                },
                {
                  "type": "null"
                }
              ],
              "title": "Family"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "threats_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Threats Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reports/pdf": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Download Pdf",
        "description": "Render the client-held analysis result without reading the uploaded sample.",
        "operationId": "download_pdf_api_v1_reports_pdf_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisResult-Input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/reports/summary": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Download Summary Pdf",
        "description": "Generate an aggregate threat-monitoring / operational report for the given window.",
        "operationId": "download_summary_pdf_api_v1_reports_summary_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "window",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(24h|7d|30d)$",
              "default": "7d",
              "title": "Window"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/history": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "List Report History",
        "operationId": "list_report_history_api_v1_reports_history_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReportRecord"
                  },
                  "title": "Response List Report History Api V1 Reports History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "List Reports",
        "description": "Every per-scan threat-prediction report generated so far, newest first.",
        "operationId": "list_reports_api_v1_reports_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReportResponse"
                  },
                  "type": "array",
                  "title": "Response List Reports Api V1 Reports Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/reports/{report_id}": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get Report Status",
        "operationId": "get_report_status_api_v1_reports__report_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/{report_id}/pdf": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Download Previous Pdf",
        "description": "Re-render the PDF for a previously generated (persisted) report.",
        "operationId": "download_previous_pdf_api_v1_reports__report_id__pdf_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalysisResult-Input": {
        "properties": {
          "object_path": {
            "type": "string",
            "title": "Object Path"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "md5": {
            "type": "string",
            "title": "Md5"
          },
          "hashes": {
            "$ref": "#/components/schemas/FileHashes"
          },
          "metadata": {
            "$ref": "#/components/schemas/FileMetadata"
          },
          "signature_match": {
            "$ref": "#/components/schemas/SignatureMatch"
          },
          "yara_matches": {
            "items": {
              "$ref": "#/components/schemas/YaraMatch"
            },
            "type": "array",
            "title": "Yara Matches",
            "default": []
          },
          "yara_available": {
            "type": "boolean",
            "title": "Yara Available",
            "default": true
          },
          "network_indicators": {
            "$ref": "#/components/schemas/NetworkIndicators",
            "default": {
              "urls": [],
              "ips": [],
              "domains": []
            }
          },
          "suspicious_indicators": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suspicious Indicators",
            "default": []
          },
          "suspicious_strings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suspicious Strings",
            "default": []
          },
          "strings_sample": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Strings Sample",
            "default": []
          },
          "risk": {
            "$ref": "#/components/schemas/RiskAssessment"
          },
          "behavior": {
            "$ref": "#/components/schemas/BehaviorProfile",
            "default": {
              "available": false,
              "is_pe": false,
              "capabilities": [],
              "narrative": "Behavioral analysis not run.",
              "pe_summary": {},
              "import_summary": {}
            }
          },
          "ml": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MLPrediction"
              },
              {
                "type": "null"
              }
            ]
          },
          "verdict": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Verdict"
              },
              {
                "type": "null"
              }
            ]
          },
          "threat_intel": {
            "$ref": "#/components/schemas/ThreatIntelResult",
            "default": {
              "configured": false,
              "available": false
            }
          },
          "behavioral_analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BehavioralAnalysisResult-Input"
              },
              {
                "type": "null"
              }
            ]
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "object_path",
          "sha256",
          "md5",
          "hashes",
          "metadata",
          "signature_match",
          "risk",
          "notes"
        ],
        "title": "AnalysisResult"
      },
...
      "FileHashes": {
        "properties": {
          "md5": {
            "type": "string",
            "title": "Md5"
          },
          "sha1": {
            "type": "string",
            "title": "Sha1"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          }
        },
        "type": "object",
        "required": [
          "md5",
          "sha1",
          "sha256"
        ],
        "title": "FileHashes",
        "description": "Cryptographic digests of the uploaded sample."
      },
      "FileMetadata": {
        "properties": {
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes",
            "description": "File size in bytes."
          },
          "file_type": {
            "type": "string",
            "title": "File Type",
            "description": "Human-readable file type from magic bytes."
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type",
            "description": "Best-effort MIME type from content."
          },
          "magic_hex": {
            "type": "string",
            "title": "Magic Hex",
            "description": "Hex of the first bytes of the file."
          },
          "extension": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extension",
            "description": "Declared file extension, lowercased (if the upload name is known)."
          },
          "extension_matches_content": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extension Matches Content",
            "description": "Whether the declared extension is consistent with the detected content type. False is a classic disguise indicator."
          },
          "shannon_entropy": {
            "type": "number",
            "title": "Shannon Entropy",
            "description": "Byte entropy in bits/byte (0-8)."
          },
          "likely_packed": {
            "type": "boolean",
            "title": "Likely Packed",
            "description": "Heuristic: binary content with high entropy, suggesting packing/encryption."
          },
          "printable_ratio": {
            "type": "number",
            "title": "Printable Ratio",
            "description": "Fraction of printable-ASCII bytes."
          },
          "likely_text": {
            "type": "boolean",
            "title": "Likely Text",
            "description": "Whether the sample looks like a text file."
          }
        },
        "type": "object",
        "required": [
          "size_bytes",
          "file_type",
          "mime_type",
          "magic_hex",
          "shannon_entropy",
          "likely_packed",
          "printable_ratio",
          "likely_text"
        ],
        "title": "FileMetadata",
        "description": "Static metadata derived from the sample without executing it."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NetworkIndicators": {
        "properties": {
          "urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Urls",
            "default": []
          },
          "ips": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Ips",
            "default": []
          },
          "domains": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Domains",
            "default": []
          }
        },
        "type": "object",
        "title": "NetworkIndicators"
      },
...
      "ReportRecord": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "report_type": {
            "type": "string",
            "title": "Report Type"
          },
          "format": {
            "type": "string",
            "title": "Format",
            "default": "pdf"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "verdict_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verdict Label"
          },
          "risk_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Score"
          },
          "window": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "report_type",
          "title"
        ],
        "title": "ReportRecord",
        "description": "One entry in the generated-report history (investigation or summary)."
      },
      "ReportResponse": {
        "properties": {
          "report_id": {
            "type": "string",
            "title": "Report Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "sample_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Id"
          },
          "file_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Hash"
          },
          "predicted_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Predicted Class"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "is_malicious": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Malicious"
          },
          "risk_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Score"
          },
          "severity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          },
          "static_indicators": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Static Indicators",
            "default": []
          },
          "recommendation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "report_id",
          "status"
        ],
        "title": "ReportResponse"
      },
      "RiskAssessment": {
        "properties": {
          "score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score",
            "description": "Aggregate static risk score, 0-100."
          },
          "level": {
            "type": "string",
            "title": "Level",
            "description": "low | medium | high"
          },
          "classification": {
            "type": "string",
            "title": "Classification",
            "description": "Human-readable verdict."
          },
          "recommended_action": {
            "type": "string",
            "title": "Recommended Action"
          }
        },
        "type": "object",
        "required": [
          "score",
          "level",
          "classification",
          "recommended_action"
        ],
        "title": "RiskAssessment"
      },
...
      "SignatureMatch": {
        "properties": {
          "matched": {
            "type": "boolean",
            "title": "Matched"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "severity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          }
        },
        "type": "object",
        "required": [
          "matched"
        ],
        "title": "SignatureMatch"
      },
...
      "UpdateProfileRequest": {
        "properties": {
          "display_name": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "title": "Display Name"
          }
        },
        "type": "object",
        "required": [
          "display_name"
        ],
        "title": "UpdateProfileRequest"
      },
      "UserProfile": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "role"
        ],
        "title": "UserProfile"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "YaraMatch": {
        "properties": {
          "rule": {
            "type": "string",
            "title": "Rule"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "default": []
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta",
            "default": {}
          },
          "matched_strings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Matched Strings",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "rule"
        ],
        "title": "YaraMatch"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}